fix(darwin::functions::lsrf): Add 1 second delay after opening a file
Opening the next file right after closing the current one caused a race condition
as open() couldn't release a lock on the application in time before opening it again,
which caused the next file to be skipped.
A 1 second delay should be enough to prevent such errors in the future.